requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cssselect2==0.7.0
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.9
django-cassandra-engine==1.8.0
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-mysql==4.12.0
django-nine==0.2.7
django-oauth-toolkit==2.3.0
django-weasyprint==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.2
fonttools==4.51.0
geomet==0.2.1.post1
gunicorn==22.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json-converter==0.5.0
jsonConverter==1.0.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.0.8
mysqlclient==2.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipenv==2023.12.1
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.9.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.6
referencing==0.35.1
reportlab==4.2.0
requests==2.31.0
rpds-py==0.19.0
scikit-learn==1.5.1
scipy==1.14.0
scylla-driver==3.26.6
seaborn==0.13.2
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
SMTPEmail==0.4.2
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
stripe==8.9.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
Werkzeug==3.0.1
xmltodict==0.12.0
xmltojson==2.0.2
yarg==0.1.9
zope.interface==6.4.post2
zopfli==0.2.3
